requirements.txt
fastapi
uvicorn
llama-cpp-python
huggingface_hub